requirements.txt
aiohttp==3.9.5
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
av==12.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.4
cryptography==42.0.8
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
distutils-pytest==0.2.1
dnspython==2.6.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.31.0
google-cloud-texttospeech==2.16.3
google-crc32c==1.5.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ifaddr==0.2.0
inflection==0.5.1
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.7
langchain-text-splitters==0.2.1
langsmith==0.1.77
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
openai==0.28.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydid==0.5.1
pydub==0.25.1
pyee==11.1.0
pygame==2.6.0
Pygments==2.18.0
pyttsx3==2.6.0
pyimgur==0.6.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pytest==8.2.2
python-dateutil==2.9.0.post0
playsound
pyalsaaudio
pyaudio
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
replicate==0.26.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
setuptools==70.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.30
stack-data==0.6.3
streamlit==1.35.0
streamlit-chat==0.1.1
streamlit-webrtc==0.47.7
streamlit-audiorecorder
tenacity==8.4.1
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
yarg==0.1.9
yarl==1.9.4
streamlit-extras
firebase-admin
google-cloud-firestore
st-pages
